BACKLOG-17140 Reorganize sections, fix ordering section, fix tests, update moonstone
