--- src/javascript/EditPanel/EditPanelContent/FormBuilder/FormBuilder.scss
@@ -1,3 +1,0 @@
-.title > button > span {
-    text-transform: uppercase;
-}
