+++ src/javascript/EditPanel/EditPanelContent/FormBuilder/Sections/ChildrenSection/AutomaticOrdering/AutomaticOrdering.jsx
