+++ src/javascript/EditPanel/EditPanelContent/FormBuilder/Sections/ChildrenSection/AutomaticOrdering/index.js
