+++ src/javascript/EditPanel/EditPanelContent/FormBuilder/Sections/ChildrenSection/ChildrenSection.spec.jsx
