+++ src/javascript/EditPanel/EditPanelContent/FormBuilder/Sections/ChildrenSection/ManualOrdering/DragDrop/DraggableReference.spec.jsx
