+++ src/javascript/EditPanel/EditPanelContent/FormBuilder/Sections/ChildrenSection/ManualOrdering/ManualOrdering.jsx
