+++ src/javascript/EditPanel/EditPanelContent/FormBuilder/Sections/ChildrenSection/ManualOrdering/index.js
