+++ src/javascript/EditPanel/EditPanelContent/FormBuilder/Sections/OrderingSection/index.js
@@ -1,0 +1,1 @@
+export * from './OrderingSection';
