+++ src/javascript/EditPanel/EditPanelContent/FormBuilder/Sections/Section/index.js
@@ -1,0 +1,1 @@
+export * from './Section';
