+++ src/javascript/EditPanel/EditPanelContent/FormBuilder/Sections/index.js
@@ -1,0 +1,3 @@
+export * from './Section';
+export * from './OrderingSection';
+export * from './ChildrenSection';
